requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
diffusers>=0.21.0
transformers>=4.30.0
hydra-core>=1.3.0
wandb>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
scikit-image>=0.21.0
kornia>=0.7.0
tifffile>=2023.0.0
cellpose>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
einops>=0.6.0
pytest>=7.3.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
jupyter>=1.0.0
ipywidgets>=8.0.0

# Added for dataset download utilities and PIL image IO
requests>=2.31.0
Pillow>=9.4.0


